move judges to pundit

+++ app/controllers/judges_controller.rb
@@ -1,14 +1,14 @@
 class JudgesController < ApplicationController
   layout "competition_management"
-  load_and_authorize_resource :competition, only: [:index, :create, :destroy, :copy_judges]
-  before_action :load_new_judge, only: [:create]
-  load_and_authorize_resource
+  before_action :load_competition, only: [:index, :create, :destroy, :copy_judges]
 
   respond_to :html
 
   # POST /competitions/#/judges
   # POST /competitions/#/judges.json
   def create
+    @judge = @competition.judges.new(judge_params)
+    authorize @judge
     if @judge.save
       flash[:notice] = 'Association was successfully created.'
     else
@@ -19,6 +19,8 @@ class JudgesController < ApplicationController
 
   # POST /event/#/judges/copy_judges
   def copy_judges
+    authorize @competition, :copy_judges?
+
     @from_event = Competition.find(params[:copy_judges][:competition_id])
     @from_event.judges.each do |source_judge|
       new_judge = @competition.judges.build
@@ -33,6 +35,9 @@ class JudgesController < ApplicationController
   # this is used to update standard_execution_scores
   # PUT /judge/1
   def update
+    @judge = Judge.find(params[:id])
+    authorize @judge
+
     respond_to do |format|
       if @judge.update_attributes(judge_params)
         format.html { redirect_to judge_standard_scores_path(@judge), notice: 'Judge Scores successfully created.' }
@@ -45,6 +50,8 @@ class JudgesController < ApplicationController
 
   # this is used to toggle the active-status of a judge
   def toggle_status
+    authorize @judge
+
     if @judge.active?
       @judge.update_attribute(:status, "removed")
     else
@@ -56,21 +63,24 @@ class JudgesController < ApplicationController
   # DELETE /judges/1
   # DELETE /judges/1.json
   def destroy
-    ec = @judge.competition
+    @judge = @competition.judges.find(params[:id])
+    authorize @judge
 
     respond_to do |format|
       if @judge.destroy
-        format.html { redirect_to competition_judges_path(ec), notice: "Judge Deleted" }
+        format.html { redirect_to competition_judges_path(@competition), notice: "Judge Deleted" }
         format.json { head :no_content }
       else
         flash[:alert] = "Unable to delete judge"
-        format.html { redirect_to competition_judges_path(ec) }
+        format.html { redirect_to competition_judges_path(@competition) }
         format.json { head :no_content }
       end
     end
   end
 
   def index
+    authorize Judge.new
+
     add_to_competition_breadcrumb(@competition)
     add_breadcrumb "Manage Judges", competition_judges_path(@competition)
 
@@ -84,8 +94,8 @@ class JudgesController < ApplicationController
 
   private
 
-  def load_new_judge
-    @judge = @competition.judges.new(judge_params)
+  def load_competition
+    @competition = Competition.find(params[:competition_id])
   end
 
   def judge_params
